apps/llm-service/pyproject.toml
[tool.poetry]
name = "llm-service"
version = "2.1.0"
description = "The Expert Analyst for NODA Copilot, powered by FastAPI and LangChain."
authors = ["NODA Team <codedhinakarr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# --- Web Framework ---
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
pydantic = "^2.7.4"

# --- Database ---
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.5"

# --- Google & LangChain Libraries ---
google-generativeai = "^0.5.4"
langchain = "0.2.5"
langchain-core = "0.2.10"
langchain-community = "0.2.5"
langchain-google-genai = "1.0.5"

# --- Utilities ---
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.5.0"
langsmith = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"